pyproject.toml
[tool.poetry]
name = "mto"
version = "0.1.0"
description = ""
authors = ["Divider <it-cio@mail.ru>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "^1.4.3"
openpyxl = "^3.0.10"
PyQt5 = "^5.15.7"
docxtpl = "^0.16.4"
xlwings = "^0.27.14"
matplotlib = "^3.5.3"
PyQt5-stubs = "^5.15.6.0"
aspose-words = "^22.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
